+++ tests/mod.rs
@@ -270,3 +270,7 @@ fn cargo_home_filtering() {
 
     assert_eq!(res.iter().count(), 1);
 }
+
+fn follow_exes_down() {
+    check_percentage("follow_exe", 1.0f64, true);
+}